bench/pyproject.toml
[project]
name = "ocm-bench"
version = "0.1.0"
description = "OpenCircuitModel benchmark + sandbox framework"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pyarrow>=15",
    "duckdb>=0.10",
    "pydantic>=2.5",
    "scipy>=1.12",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "tomli>=2.0; python_version<'3.11'",
    "tomli-w>=1.0",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[project.scripts]
bench = "bench.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["bench*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "UP"]
